chore: 🚚 Move utility files into appropriate folder
Moved utility files into folder named the same. Also consolidated utility methods into single file. Updated all imports and references.
