+++ nautobot_ssot_infoblox/diffsync/adapters/infoblox.py
@@ -5,7 +5,7 @@ import re
 from diffsync import DiffSync
 from diffsync.enum import DiffSyncFlags
 from nautobot.extras.plugins.exceptions import PluginImproperlyConfigured
-from nautobot_ssot_infoblox.diffsync.client import InfobloxApi
+from nautobot_ssot_infoblox.utils.client import InfobloxApi
 from nautobot_ssot_infoblox.diffsync.models.infoblox import (
     InfobloxAggregate,
     InfobloxIPAddress,
