+++ nautobot_ssot_infoblox/diffsync/models/nautobot.py
@@ -8,7 +8,7 @@ from nautobot.ipam.models import Prefix as OrmPrefix
 from nautobot.ipam.models import VLAN as OrmVlan
 from nautobot.ipam.models import VLANGroup as OrmVlanGroup
 from nautobot_ssot_infoblox.diffsync.models.base import Aggregate, Network, IPAddress, Vlan, VlanView
-from nautobot_ssot_infoblox.diffsync.utilities import create_tag_sync_from_infoblox
+from nautobot_ssot_infoblox.utils.utilities import create_tag_sync_from_infoblox
 
 
 class NautobotNetwork(Network):
